pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Abraão Zaidan <abraao.zaidan@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
tda-api = "^1.6.0"
pyqt6 = "^6.6.0"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
